% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{cppd.check_probes}
\alias{cppd.check_probes}
\title{QC test pbat capture probes}
\usage{
cppd.check_probes(probes, conf_fn, defaults_fn = NULL, log_fn = NULL)
}
\arguments{
\item{probes}{data frame with probes (output of cppd.generate_probes)}

\item{conf_fn}{configuration file}

\item{defaults_fn}{default configuration file (optional)}

\item{log_fn}{log file (optional)}
}
\description{
QC test pbat capture probes
}
