% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{cppd.create_genome_kmer_db}
\alias{cppd.create_genome_kmer_db}
\title{Generate kmers db}
\usage{
cppd.create_genome_kmer_db(db_ofn, jellyfish_bin, jellyfish_params = "",
  k = 15, fasta_files = NULL, hash_size = round(memfree()/3),
  threads = getOption("cppd.parallel.thread_num"))
}
\arguments{
\item{db_ofn}{prefix of kmers database files}

\item{jellyfish_bin}{jellyfish binary}

\item{jellyfish_params}{parameters to jellyfish}

\item{k}{kmer size}

\item{fasta_files}{genomic fasta files to use if there are no precomputed converted and unconverted fasta files ('conv_methylated.fasta' and 'conv_unmethylated.fasta')}

\item{hash_size}{jellyfish -s paramter}

\item{threads}{jellyfish threads parameter (-t)}
}
\description{
Generate kmers db
}
