% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cppd.dump_example_config}
\alias{cppd.dump_example_config}
\title{Dump config file templates.}
\usage{
cppd.dump_example_config(path)
}
\arguments{
\item{path}{directory to dump files to.}
}
\description{
Dump templates of config files required by the package.
}
