% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{cppd.gene_enhancers}
\alias{cppd.gene_enhancers}
\title{Filter for putative enhancers that are within gene tads}
\usage{
cppd.gene_enhancers(genes, intervals, tads)
}
\arguments{
\item{genes}{character vector of gene symbols}

\item{intervals}{putative enhancer intervals}

\item{tads}{tad intervals: intervals set with an extra id column}
}
\value{
intervals from \code{intervals} with an extra 'id' column for TAD id and 'gene' with the genes from \code{genes} within the TAD
}
\description{
Filter for putative enhancers that are within gene tads
}
