% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{cppd.gene_promoters}
\alias{cppd.gene_promoters}
\title{Get promoter region of genes}
\usage{
cppd.gene_promoters(genes, upstream = 500, downstream = 50)
}
\arguments{
\item{genes}{character vector of gene symbols}

\item{upstream}{number of base pairs upstream to the gene}

\item{downstream}{number of base pairs downstream to the gene}
}
\value{
intervals set with the gene promoters
}
\description{
Get promoter region of genes
}
