% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cppd.generate_probes}
\alias{cppd.generate_probes}
\title{Generate pbat capture probes}
\usage{
cppd.generate_probes(conf_fn, defaults_fn = NULL, log_fn = NULL,
  return_probes = FALSE)
}
\arguments{
\item{conf_fn}{configuration file}

\item{defaults_fn}{default configuration file (optional)}

\item{log_fn}{log file (optional)}

\item{return_probes}{return a data frame with the probes}
}
\value{
probes data frame if return_probes is TRUE, and NULL otherwise
}
\description{
Generate pbat capture probes
}
