% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{cppd.intervals_cpgs}
\alias{cppd.intervals_cpgs}
\title{Get CpGs within a region}
\usage{
cppd.intervals_cpgs(intervals)
}
\arguments{
\item{intervals}{intervals set}
}
\value{
intervals set of CpGs within the given intervals
}
\description{
Get CpGs within a region
}
