% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cppd.set_parallel}
\alias{cppd.set_parallel}
\title{Set parallel threads}
\usage{
cppd.set_parallel(thread_num)
}
\arguments{
\item{thread_num}{number of threads. use '1' for non parallel behaviour}
}
\value{
None
}
\description{
Set parallel threads
}
\examples{
cppd.set_parallel(8)

}
